+++ RELEASES.md
@@ -4,6 +4,7 @@ Released on TBD (UTC).
 
 ### Enhancements
 - Added streaming methods for `TardisCSVDataLoader`
+- Added stream iterators support for `BacktestEngine` low-level streaming API
 - Added `YEAR` aggregation and improved bar specification validation (#2771), thanks @stastnypremysl
 - Added support for requesting any number of historical bars for dYdX (#2766, #2777), thanks @DeirhX
 - Added `use_hyphens_in_client_order_ids` config option for `StrategyConfig`
@@ -15,6 +16,7 @@ Released on TBD (UTC).
 
 ### Internal Improvements
 - Refactored OKX adapter to Rust API clients
+- Added stream iterators support `BacktestDataIterator`
 - Added `start` and `stop` to response data (#2748), thanks @stastnypremysl
 - Added integration test service management targets (#2765), thanks @stastnypremysl
 - Added integration tests for dYdX bar-partitioning and large-history handling (#2773), thanks @nicolad